backend/requirements.txt
# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
scrapy==2.11.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Vector database
pinecone-client==2.2.4
weaviate-client==3.25.3
faiss-cpu==1.7.4

# Machine Learning / Embeddings
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
openai==1.3.5

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
celery==5.3.4
redis==5.0.1
pyyaml==6.0.1
types-PyYAML==6.0.12.12

# Logging and monitoring
loguru==0.7.2

# Language processing
langdetect==1.0.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
pytest-html==4.1.1
factory-boy==3.3.0
faker==20.1.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0
bandit==1.7.5
